extra color for own notes
