+++ src/app/wine-tastings/wine/wine.component.scss
@@ -11,6 +11,7 @@
   background: url('../../../assets/icons/down-arrow.png') center bottom no-repeat;
   background-size: 24px;
   border-bottom: 1px solid #cdcdcd;
+  cursor: pointer;
 
   &.allRatings--opened {
     padding-bottom: 0.5rem;
@@ -39,27 +40,12 @@
   margin-bottom: 0.5rem;
   padding: 0.5rem;
 
+  &.note--mine {
+    border-bottom: 2px solid #673ab7;
+  }
+
   .details {
+    opacity: 0.8;
     text-align: right;
   }
-}
-
-.spn-add-wine-notes {
-  width: 100%;
-  display: inline-block;
-}
-
-.button-wrapper {
-  display: inline-block;
-}
-
-.btn-rating-edit {
-  padding: 0 0 3px 0;
-  margin-left: 5px;
-  background: url('../../../assets/icons/edit_17.svg') no-repeat;
-  background-size: 14px;
-  border: none;
-  width: 14px;
-  height: 14px;
-  cursor: pointer;
 }
